@@ -6,17 +6,17 @@
     <packaging>jar</packaging>
     <groupId>edu.isi.oba</groupId>
     <artifactId>oba</artifactId>
-    <version>3.6.1</version>
+    <version>3.7.0</version>
     <name>core</name>
     <url>https://github.com/KnowledgeCaptureAndDiscovery/OBA</url>
 
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <jdk.version>1.8</jdk.version>
-        <maven-shade-plugin.version>3.1.1</maven-shade-plugin.version>
+        <maven.compiler.release>11</maven.compiler.release>
+        <jdk.version>11</jdk.version>
+        <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-shade-plugin -->
+        <maven-shade-plugin.version>3.5.2</maven-shade-plugin.version>
     </properties>
 
     <dependencies>
