@@ -2,24 +2,26 @@ package edu.isi.oba;
 
 import static edu.isi.oba.ObaUtils.get_yaml_data;
 import edu.isi.oba.config.YamlConfig;
-import java.io.File;
-import org.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Test;
 
+import java.io.File;
 import java.io.IOException;
 
-import static org.junit.Assert.*;
+import org.json.JSONObject;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
 
 import org.semanticweb.owlapi.model.OWLClass;
 import org.semanticweb.owlapi.model.OWLOntologyCreationException;
 
+@Disabled
 public class ObaUtilsTest {
 
     @Test
     public void read_json_file() throws IOException {
         JSONObject actual = ObaUtils.read_json_file("json_one.json");
-        Assert.assertNotNull(actual.get("@context"));
+        Assertions.assertNotNull(actual.get("@context"));
     }
 
     @Test
@@ -27,8 +29,8 @@ public class ObaUtilsTest {
         JSONObject one = ObaUtils.read_json_file("json_one.json");
         JSONObject two = ObaUtils.read_json_file("json_two.json");
         JSONObject merge = ObaUtils.mergeJSONObjects(one, two);
-        Assert.assertNotNull(merge.get("@context"));
-        Assert.assertNotNull(merge.get("@context"));
+        Assertions.assertNotNull(merge.get("@context"));
+        Assertions.assertNotNull(merge.get("@context"));
     }
 
     @Test
@@ -39,9 +41,9 @@ public class ObaUtilsTest {
         JSONObject[] jsons = new JSONObject[]{ one, two, three};
         JSONObject merge = ObaUtils.concat_json_common_key(jsons, "@context");
         JSONObject o = (JSONObject) merge.get("@context");
-        assertNotNull(o.get("Entity"));
-        assertNotNull(o.get("Model"));
-        assertNotNull(o.get("Setup"));
+        Assertions.assertNotNull(o.get("Entity"));
+        Assertions.assertNotNull(o.get("Model"));
+        Assertions.assertNotNull(o.get("Setup"));
     }
     
     @Test
@@ -52,9 +54,9 @@ public class ObaUtilsTest {
             Mapper mapper = new Mapper(config_data);
             OWLClass planClass = mapper.manager.getOWLDataFactory().getOWLClass("http://purl.org/net/p-plan#Plan");
             String desc = ObaUtils.getDescription(planClass, mapper.ontologies.get(0));
-            assertNotEquals(desc, "");
+            Assertions.assertNotEquals(desc, "");
         }catch(Exception e){
-            fail();
+            Assertions.fail("Failed to get description.", e);
         }
     }
 
@@ -69,8 +71,7 @@ public class ObaUtilsTest {
         YamlConfig config_data = get_yaml_data(missing_file);
         try {
             Mapper mapper = new Mapper(config_data);
-            //if no exception is launched, fail test
-            fail();
+            Assertions.fail("Missing file: If no exception is launched, fail test");
         }catch(Exception e){
             //pass test if there is an exception
         }
@@ -91,17 +92,17 @@ public class ObaUtilsTest {
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         JSONObject o = (JSONObject) context.get("@context");
-        assertEquals(o.get("id"), "@id");
-        assertEquals(o.get("type"), "@type");
-        assertNotNull(o.get("Entity"));
-        assertNotNull(o.get("Model"));
+        Assertions.assertEquals(o.get("id"), "@id");
+        Assertions.assertEquals(o.get("type"), "@type");
+        Assertions.assertNotNull(o.get("Entity"));
+        Assertions.assertNotNull(o.get("Model"));
         try{
             java.nio.file.Files.delete(ont1.toPath());
             java.nio.file.Files.delete(ont2.toPath());
         }catch(Exception e){
         }
-
     }
 
 }
