@@ -1,13 +1,13 @@
 package edu.isi.oba.config;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class ProviderTest {
     @Test
     public void getSelectedClasses() {
         Provider provider = Provider.get("firebase");
         Provider provider_test = Provider.FIREBASE;
-        Assert.assertEquals(provider_test, provider);
+        Assertions.assertEquals(provider_test, provider);
     }
 }
