@@ -1,13 +1,12 @@
 package edu.isi.oba.config;
 
-import org.junit.Assert;
-import org.junit.Test;
+import static edu.isi.oba.ObaUtils.get_yaml_data;
 
 import java.util.Arrays;
 import java.util.List;
 
-import static edu.isi.oba.ObaUtils.get_yaml_data;
-import static org.junit.Assert.assertThat;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class YamlConfigTest {
     @Test
@@ -16,6 +15,6 @@ public class YamlConfigTest {
         YamlConfig config_data = get_yaml_data(config_test_file_path);
         List<String> expected = Arrays.asList("http://dbpedia.org/ontology/Genre", "http://dbpedia.org/ontology/Band");
         List<String> config = config_data.getClasses();
-        Assert.assertEquals(expected, config);
+        Assertions.assertEquals(expected, config);
     }
 }
